Draw multiple tiles from world

+++ src/tile.rs
@@ -10,10 +10,10 @@ pub enum Tile {
 
 impl Tile {
     /// A tile's width in pixels.
-    pub const WIDTH: i32 = 16;
+    pub const WIDTH: usize = 16;
 
     /// A tile's height in pixels.
-    pub const HEIGHT: i32 = Self::WIDTH;
+    pub const HEIGHT: usize = Self::WIDTH;
 
     /// Returns the tile's color.
     pub fn get_color(self) -> u32 {
